move most of the logic relating to the search bar to the header component's index.js
this makes more sense. the screenController should just be importing most things and orchestrating the entire screen. the display() function certainly does this
